src/main/java/de/timmi6790/mineplexleaderboardupdate/utilities/UUIDUtilities.java
package de.timmi6790.mineplexleaderboardupdate.utilities;

import lombok.experimental.UtilityClass;

import java.nio.ByteBuffer;
import java.util.UUID;

@UtilityClass
public class UUIDUtilities {
    public byte[] getBytesFromUUID(final UUID uuid) {
        return ByteBuffer.wrap(new byte[16])
                .putLong(uuid.getMostSignificantBits())
                .putLong(uuid.getLeastSignificantBits())
                .array();
    }

    public UUID getUUIDFromBytes(final byte[] bytes) {
        final ByteBuffer byteBuffer = ByteBuffer.wrap(bytes);
        return new UUID(byteBuffer.getLong(), byteBuffer.getLong());
    }
}
